Revert "Remove yacss's poll PMT patch"
This reverts commit 165dd54c082b300a851fbce41f05c24aace27e7f.

The benefits of this patch are to big to ditch it.
For one thing: it avoids a total lock-up of a single mux in case one of the scheduled shows stops recording.
But that is just masquing a bigger problem: the tuning mechanism needs some (a lot of) work.
For now: use the patch.